move push operation, but not push32
